\name{Chicago-package}
\alias{Chicago-package}
\alias{Chicago}
\docType{package}
\title{
\packageTitle{Chicago}
}
\description{
\packageDescription{Chicago}
}
\details{
To get started, please read the vignette: \code{vignette("Chicago")}

The DESCRIPTION file:
\packageDESCRIPTION{Chicago}
\packageIndices{Chicago}
}
\author{
\packageAuthor{Chicago}

Maintainer: \packageMaintainer{Chicago}
}
\references{

}
\keyword{ package }

