\name{Chicago-package}
\alias{Chicago-package}
\alias{Chicago}
\docType{package}
\title{
CHiCAGO: Capture Hi-C Analysis of Genomic Organization
}
\description{
A pipeline for analysing Capture Hi-C data.
}
\details{
To get started, please read the vignette: \code{vignette("Chicago")}

%The DESCRIPTION file:
%\packageDESCRIPTION{Chicago}
%\packageIndices{Chicago}
}
\author{
Jonathan Cairns, Paula Freire Pritchett, Steven Wingett, Mikhail Spivakov

Maintainer: Mikhail Spivakov - \email{spivakov@babraham.ac.uk}
}
\keyword{ package }

